added command to display all the files in current directory #2
